fix the failing RetryAfter test
The test was failing because some values of the timestamp don't make
it into the formatted representation: the nanoseconds and the
year-day. We also currently convert all timestamps to UTC before
formatting. Supply a timestamp to the tests which has zero for nsec
and yday, and is already in UTC; straight equality testing then works.